requirements.txt
pandas
numpy
newsapi-python
yfinance
nltk
spacy
matplotlib
seaborn
plotly
python-dotenv
pytest
pytest-mock
PyYAML
